tracer: refine unit test
